Task: paircalc section proxy cleanup
Summary: Cleaned up the section proxy storage and manipulation in class PairCalcID and related functions.  There are now simply two array sections in PairCalcID (one each for left and right data destinations) irrespective of whether the PC array is for symm or asymm loops. Section names are sectionGettingLeft and sectionGettingRight.

Also cleaned up the make*Tree() and startPairCalc*() functions. The methods now look remarkably similar except for the fact that they operate on the left and right sections respectively. Its not difficult to envision some means of collapsing these two into one, but there seems to be no need for it right now.

Percentage: 95%
Hours: 15