Task: remove acceptPhantomData Summary: PairCalculator::acceptPhantomData() has been removed.
Phantom chares now get data via acceptRightData() in the same msg type as non-phantoms (paircalcInputMsg). The computation trigger logic for all of the forward path is now consolidated in launchComputations(). Phantom chares also compute directly on the data in the message. Tests run fine for min/dyn and conserveMem=0/1 with phantoms on or off. This task is done, but there are a couple of follow-ups:

- confirm that there are no message leaks in any of the paths (phantom/non-phantom, symm/asymm, norma/psiv loops) for any conserveMemory setting, and that msgLeft and msgRight get deleted as appropriate.
- Setting flag_dp = true/false in the code does not function unversally for min/dynamics. Look into this and confirm that doublepack settings are done right for min/dyn.

Progress: 100%
Hours: 16