Task: remove acceptPhantomData Summary: numRecd is not reset in launchComputations() if forward path will be triggered later.
When gamma beats OrthoT and arrives first, the forward path cannot be launched until T arrives. However, launchComputations() was resetting numRecd, so there was no way to figure out if we were ready to launch forward path in acceptOrthoT() even when T arrived completely. Now, launchComputations checks for this race before resetting numRecd, and if the race does happen, numRecd is reset in acceptOrthoT(). This kind of breaks the efforts to localize numRecd to just the acceptData and launchComputation methods, but cannot be helped.

Progress: 100%
Hours: 2