Knocked out all the explicit arrays of planes in the simulation being passed around. Cleans up after commit 6794284d91e6386a84cdcab0a0406b07910863dc
Chare instantiation used this list of planes to include which made sense in early versions of the code because of some logic associated with paircalc instantiation. This commit from 4 years ago removed that logic but did not cleanup after itself. Removed this trail of redundant arrays of planes being passed around during the instantiation phase.