cp::ortho::ArrayBuilder::build() replaces init_ortho_chares()
The intent is to decouple ortho mapping and array creation code from the
mess of globals, readonlys and other dependencies in cpaimd.C so that it
can be used independently to spawn ortho (and related) arrays. Currently,
this is just migration of code to another file/class. More cleanup
and decoupling is required and will follow.