Fix bug in make dependency generation rules.
Escape a period in the sed search pattern expression to ensure it is not interpreted as 'any' character.
This was causing build issues under specific sets of file names. For eg, if bar.o depended on barConfig.h,
barCo would also match the pattern bar.o, and spurious substitutions would occur. However if the search pattern
is bar\.o, the issue is avoided.