Fix bug in paircalc grainsize remainder logic
In the ancient (~20 months ago) days of paircalc history, there was a method acceptPhantomData
that would allow the phantoms to accept incoming data from their non-phantom mirror chares.

Phantoms were sent right matrix data by their mirrors and they saved it in the inDataRight buffer
and operated on it as right matrix data. The current scheme where acceptPhantomData has been
consumed by acceptRightData() and launchComputations() preserves all of this process.

However, there was one catch in acceptPhantomData. Even though the phantom received data as a right
matrix and saved it in the right matrix buffer, the size of the buffer was actually = the size of
the left matrix buffer that the phantom would have had were it not a phantom! This makes sense if
we pause and think a second: The right matrix block (Bj) for a non-phantom mirror is
actually the left matrix block for the phantom chare (Aj).

This tiny catch got lost during the upheaval paircalc has gone through and this commit fixes that
in by simply swapping the computed left and right matrix block sizes for the phantoms alone. Not
clean, but preserves the wisdom of the ancients.

Ideally, the phantoms should receive the block as a left matrix buffer, compute on it accordingly
and send it back to GSpace. But making that happen has a non-trivial prerequisite:
unraveling the phantom control/data flow thread in the BW path from the rest of the BW path logic