Start distinguishing between input and internal data types in the paircalc world
GSpace always sends paircalcs complex data. Depending on the intent of the simulation (k-points etc)
the paircalcs are expected to either perform GEMMs on real numbers or complex numbers.
This means that the data type used to handle data, perform gemms, and allocate buffers
needs to be independently chosen (double/complex). Clarify this with a new typedef.

More work needs to be done to distinguise between inputType and internalType through the paircalc/ortho code.