WIP: Introdce internalType into code that was overlooked earlier
The S matrix generated in the paircalc fw path was received as a double matrix
instead of internalType. A routine to calculate the max diagonal element in this
matrix was also only friendly to doubles. Fix these to process complex correctly too.

Had overlooked a factor of 2 (pcDataSizeFactor) in the bw path. Fixed this too