Add scripts to convert datasets to new directories
To convert a data set, copy "convert" to the data set
folder and run it.

Run "setup" in the run directory of OpenAtom to create
the necessary output folders.  The folder must have
STATES and ATOM_COORDS_IN folders (or symlinks)
present.

Use "tidy" in the run directory to clean up files
before running OpenAtom again.