Fix NaN "bug" caused by IBM compiler picking wrong abs() implementation
Complex number capable paircalc requires the use of C++ math routines that are
overloaded for complex numbers too. Hence, I replaced fabs() with abs() in code
to compute max residual values.

However, the C abs() returns an integer as opposed to the C++ abs() which
should return the same type as the input (double). Because of IBM's compiler
implementation, only the C abs() is available in the global namespace and
hence, all residual calculations in Ortho were getting truncated to integers!
Other compilers on other platforms behave differently, and hence, the issue is
only visible on the BGL and BGP.

The fix is to prefix abs() with std:: to clarify that the appropriate
overloaded abs() which should always be present in namespace std:: should be
used.