paircalc: use complex conjugate instead of just transpose in backward gemm
Akin to the forward path, the backward path also requires a complex conjugate
instead of just transpose when operating on complex matrices.

Every time Glenn visits he realizes and conveys the need for a conjugate in some
paircalc gemm. We've fixed two so far in fwd and bkwd paths. Perhaps there's one
more lurking omission in the bkwd for dynamics thats waiting for his next visit.