Substep timing: Measure RealSpaceForward phase from the arrival of the first RSFFTMsg
CP_State_RealSpacePlane::acceptFFT() was contributing to the phase start timer redn
every time it received an FFTMsg. These multiple reductions would have overwritten
the start time of the phase so that we'd actually be measuring the phase
from the first arrival of the last FFTMsg on any RealSpaceChares. This is not a
sensible number. Instead, the phase is now measured from the arrival of the first ever
FFTMsg from GSpace.