bugfix: correctness issue streaming output of symmetric backward path
The index for the completed column was the transpose of the correct value.
That created a race condition wherein some results were sent before all
contributions had been computed.  This version fixes that bug so that
streaming output produces identical output to non-streaming, and fully
supports the remainder logic cases with the same validity.