Add Matrix as a base class to PMatrix
Needs more cleanup, but this gets matrix funcionality into the PMatrix
and unifies the two versions we had to one. Uses matrixCopy to tranfer
between decompositions. A basic transpose is still needed in the matrix
base class before that can be transferred away from PMatrix. Similiarly
a copy-with-cutoff may be useful?

Change-Id: Ie9252d1c82f049d71f264b6afb17a4b5be80a747